refactor: delete disabled Google Calendar sync and collapse oauth-state
Remove the entire Google Calendar integration (route, factory, mapper,
types, constants) which was permanently disabled via ServiceUnavailableError.
Collapse oauth-state into a single file, removing unused types and helpers.

Changes:
- Delete src/lib/integrations/google-calendar/ (5 files, ~426 LOC)
- Delete API route src/app/api/v1/integrations/google-calendar/sync/
- Delete 6 related test files (unit, integration, e2e)
- Merge oauth-state-store.ts + oauth-state.types.ts into oauth-state.ts
- Remove unused IssueOAuthStateParams, ConsumeOAuthStateParams, cleanupExpiredTokens
- Update integrations AGENTS.md and related docs
