+++ src/lib/integrations/AGENTS.md
@@ -4,84 +4,20 @@
 
 ## Overview
 
-Third-party sync for Google Calendar. Uses dependency injection for testability.
+Shared OAuth utilities and CSRF state management for third-party integrations.
 
-## Architecture (DI Pattern)
-
-Each integration follows factory + types + sync structure:
+## Structure
 
 ```
 integrations/
-├── oauth.ts              # Shared OAuth utilities
-├── oauth-state.ts        # CSRF state management
-└── google-calendar/
-    ├── types.ts          # GoogleCalendarClient interface
-    ├── factory.ts        # createGoogleCalendarClient()
-    ├── sync.ts           # syncPlanToGoogleCalendar()
-    └── mapper.ts         # Tasks → Calendar events
-```
-
-## Key Pattern
-
-**Sync functions accept minimal interfaces, not SDK types:**
-
-```typescript
-// types.ts - minimal interface
-export interface GoogleCalendarClient {
-  events: {
-    insert(params: { calendarId: string; requestBody: Event }): Promise<{ data: Event }>;
-    delete(params: { calendarId: string; eventId: string }): Promise<void>;
-  };
-}
-
-// sync.ts - depends on interface, not SDK
-export async function syncPlanToGoogleCalendar(
-  planId: string,
-  client: GoogleCalendarClient  // Injected, not imported
-): Promise<void> { ... }
-
-// factory.ts - creates real client at API boundary
-export function createGoogleCalendarClient(tokens: GoogleTokens): GoogleCalendarClient {
-  const oauth2Client = new google.auth.OAuth2(...);
-  // Returns object matching interface
-}
-```
-
-## API Routes
-
-```typescript
-// src/app/api/v1/integrations/google-calendar/sync/route.ts
-import { createGoogleCalendarClient } from '@/lib/integrations/google-calendar/factory';
-import { syncPlanToGoogleCalendar } from '@/lib/integrations/google-calendar/sync';
-
-export async function POST(request: Request) {
-  const tokens = await getTokensFromDb(userId);
-  const client = createGoogleCalendarClient(tokens); // Factory at boundary
-  await syncPlanToGoogleCalendar(planId, client);
-}
-```
-
-## Testing
-
-Pass mock clients directly - no module mocking needed:
-
-```typescript
-const mockClient: GoogleCalendarClient = {
-  events: {
-    insert: vi.fn().mockResolvedValue({ data: { id: 'event_1' } }),
-    delete: vi.fn().mockResolvedValue(undefined),
-  },
-};
-
-await syncPlanToGoogleCalendar(planId, mockClient);
-expect(mockClient.events.insert).toHaveBeenCalledWith(...);
+├── oauth.ts          # Token encrypt/decrypt, store/retrieve/revoke
+└── oauth-state.ts    # CSRF state token issue/consume (single-file module)
 ```
 
 ## OAuth Flow
 
 1. **Initiate**: `/api/v1/auth/google` → redirect to Google
 2. **Callback**: `/api/v1/auth/google/callback` → exchange code, store tokens
-3. **Sync**: `/api/v1/integrations/google-calendar/sync` → use stored tokens
 
 Tokens stored in `integration_tokens` table with `provider` enum.
 
@@ -90,12 +26,10 @@ Tokens stored in `integration_tokens` table with `provider` enum.
 Credentials via `@/lib/config/env`:
 
 - `googleOAuthEnv.clientId`, `.clientSecret`, `.redirectUri`
+- `oauthEncryptionEnv.encryptionKey`
 
 Prod-required, test-optional semantics.
 
 ## Anti-Patterns
 
-- Importing SDK types in sync functions (use minimal interfaces)
-- Creating clients inside sync functions (use factory at boundary)
-- Module mocking instead of DI
 - Hardcoding credentials (use `@/lib/config/env`)
