--- src/lib/integrations/google-calendar/constants.ts
@@ -1,5 +1,0 @@
-/** Maximum number of retry attempts for Google Calendar API calls. */
-export const GOOGLE_CALENDAR_MAX_RETRIES = 3;
-
-/** Base delay in milliseconds for exponential backoff between retries. */
-export const GOOGLE_CALENDAR_RETRY_BASE_MS = 1000;
