--- src/lib/integrations/google-calendar/sync.ts
@@ -1,268 +1,0 @@
-import { getDb } from '@/lib/db/runtime';
-import {
-  googleCalendarSyncState,
-  learningPlans,
-  modules,
-  taskCalendarEvents,
-  tasks,
-} from '@/lib/db/schema';
-import { eq, inArray } from 'drizzle-orm';
-
-import {
-  AppError,
-  IntegrationSyncError,
-  NotFoundError,
-  ValidationError,
-} from '@/lib/api/errors';
-
-import { logger } from '@/lib/logging/logger';
-
-import {
-  GOOGLE_CALENDAR_MAX_RETRIES,
-  GOOGLE_CALENDAR_RETRY_BASE_MS,
-} from './constants';
-import { generateSchedule, mapTaskToCalendarEvent } from './mapper';
-import type { GoogleCalendarClient } from './types';
-
-type TaskSyncError = {
-  taskId: string;
-  message: string;
-  code?: string;
-  details?: unknown;
-};
-
-/**
- * Syncs plan tasks to Google Calendar using a pre-configured client.
- * This function is pure with respect to env and third-party client construction.
- */
-export async function syncPlanToGoogleCalendar(
-  planId: string,
-  calendarClient: GoogleCalendarClient
-): Promise<number> {
-  // Fetch plan data (RLS-enforced via getDb)
-  const db = getDb();
-  const [plan] = await db
-    .select()
-    .from(learningPlans)
-    .where(eq(learningPlans.id, planId))
-    .limit(1);
-
-  if (!plan) {
-    throw new NotFoundError('Plan not found');
-  }
-
-  // Fetch modules ordered by their order field
-  const planModules = await db
-    .select()
-    .from(modules)
-    .where(eq(modules.planId, planId))
-    .orderBy(modules.order);
-
-  if (planModules.length === 0) {
-    throw new ValidationError('No modules found for plan');
-  }
-
-  // Fetch tasks for these modules using efficient DB filtering
-  const moduleIds = planModules.map((m) => m.id);
-  const allTasks = await db
-    .select()
-    .from(tasks)
-    .where(inArray(tasks.moduleId, moduleIds));
-
-  if (allTasks.length === 0) {
-    throw new ValidationError('No tasks found for plan');
-  }
-
-  // Build module index map from planModules to preserve plan sequence
-  const moduleIndex = new Map<string, number>();
-  for (let i = 0; i < planModules.length; i++) {
-    moduleIndex.set(planModules[i].id, i);
-  }
-
-  // Stable sort tasks by: module order, then task.order, then createdAt, then id
-  const sortedTasks = allTasks.toSorted((a, b) => {
-    const aModuleIdx = moduleIndex.get(a.moduleId) ?? Number.POSITIVE_INFINITY;
-    const bModuleIdx = moduleIndex.get(b.moduleId) ?? Number.POSITIVE_INFINITY;
-    if (aModuleIdx !== bModuleIdx) return aModuleIdx - bModuleIdx;
-
-    const aOrder = typeof a.order === 'number' ? a.order : null;
-    const bOrder = typeof b.order === 'number' ? b.order : null;
-    if (aOrder !== null || bOrder !== null) {
-      const ao = aOrder ?? Number.POSITIVE_INFINITY;
-      const bo = bOrder ?? Number.POSITIVE_INFINITY;
-      if (ao !== bo) return ao - bo;
-    }
-
-    const toTime = (d: unknown): number | null => {
-      if (!d) return null;
-      if (d instanceof Date) return d.getTime();
-      if (typeof d === 'string' || typeof d === 'number') {
-        const t = new Date(d).getTime();
-        return Number.isNaN(t) ? null : t;
-      }
-      return null;
-    };
-    const aTime = toTime(a.createdAt);
-    const bTime = toTime(b.createdAt);
-    if (aTime !== null || bTime !== null) {
-      const at = aTime ?? Number.POSITIVE_INFINITY;
-      const bt = bTime ?? Number.POSITIVE_INFINITY;
-      if (at !== bt) return at - bt;
-    }
-
-    return String(a.id).localeCompare(String(b.id));
-  });
-
-  // Map tasks to the format expected by generateSchedule
-  const mappedTasks = sortedTasks.map((task) => ({
-    id: task.id,
-    title: task.title,
-    description: task.description,
-    estimatedMinutes: task.estimatedMinutes,
-  }));
-
-  // Generate schedule
-  const schedule = generateSchedule(mappedTasks, plan.weeklyHours);
-
-  if (schedule.size === 0) {
-    throw new ValidationError('Unable to schedule tasks for plan');
-  }
-
-  let eventsCreated = 0;
-  const taskErrors: TaskSyncError[] = [];
-
-  for (const task of sortedTasks) {
-    const startTime = schedule.get(task.id);
-    if (!startTime) continue;
-
-    try {
-      // Check if event already exists for this task
-      const [existingMapping] = await db
-        .select()
-        .from(taskCalendarEvents)
-        .where(eq(taskCalendarEvents.taskId, task.id))
-        .limit(1);
-
-      if (existingMapping) {
-        // Event already exists, skip
-        continue;
-      }
-
-      // Map task to calendar event format
-      const eventData = mapTaskToCalendarEvent(
-        {
-          id: task.id,
-          title: task.title,
-          description: task.description,
-          estimatedMinutes: task.estimatedMinutes,
-        },
-        startTime
-      );
-
-      // Create event in Google Calendar with retry logic
-      let event;
-      let retries = GOOGLE_CALENDAR_MAX_RETRIES;
-      while (retries > 0) {
-        try {
-          const response = await calendarClient.events.insert({
-            calendarId: 'primary',
-            requestBody: eventData,
-          });
-          event = response.data;
-          break;
-        } catch (apiError) {
-          retries--;
-          if (retries === 0) throw apiError;
-          // Wait before retry (exponential backoff)
-          await new Promise((resolve) =>
-            setTimeout(
-              resolve,
-              GOOGLE_CALENDAR_RETRY_BASE_MS *
-                (GOOGLE_CALENDAR_MAX_RETRIES + 1 - retries)
-            )
-          );
-        }
-      }
-
-      // Validate event ID exists
-      if (!event?.id) {
-        throw new Error('Google Calendar did not return a valid event ID');
-      }
-
-      // Store mapping between task and calendar event
-      try {
-        await db.insert(taskCalendarEvents).values({
-          taskId: task.id,
-          userId: plan.userId,
-          calendarEventId: event.id,
-          calendarId: 'primary',
-        });
-        eventsCreated++;
-      } catch (dbError) {
-        // DB insert failed, delete the created event to avoid orphans
-        try {
-          await calendarClient.events.delete({
-            calendarId: 'primary',
-            eventId: event.id,
-          });
-        } catch (cleanupError) {
-          logger.warn(
-            {
-              err: cleanupError,
-              calendarEventId: event.id,
-              taskId: task.id,
-            },
-            'Failed to delete orphaned calendar event after DB insert failure'
-          );
-        }
-        throw dbError;
-      }
-    } catch (error) {
-      // Record per-task error but continue syncing remaining tasks.
-      const message = error instanceof Error ? error.message : 'Unknown error';
-      const taskError: TaskSyncError = {
-        taskId: task.id,
-        message,
-      };
-      if (error instanceof AppError) {
-        const code = error.code();
-        const details = error.details();
-        if (code) {
-          taskError.code = code;
-        }
-        if (details !== undefined) {
-          taskError.details = details;
-        }
-      }
-      taskErrors.push(taskError);
-    }
-  }
-
-  // If all tasks failed, surface a structured integration error with details.
-  if (taskErrors.length > 0 && eventsCreated === 0) {
-    throw new IntegrationSyncError('Google Calendar sync failed', {
-      taskErrors,
-    });
-  }
-
-  // Store sync state
-  const syncState = {
-    planId,
-    userId: plan.userId,
-    calendarId: 'primary',
-    lastSyncedAt: new Date(),
-  };
-
-  await db
-    .insert(googleCalendarSyncState)
-    .values(syncState)
-    .onConflictDoUpdate({
-      target: googleCalendarSyncState.planId,
-      set: {
-        ...syncState,
-        updatedAt: new Date(),
-      },
-    });
-
-  return eventsCreated;
-}
