+++ src/lib/integrations/oauth-state.ts
@@ -1,5 +1,62 @@
-import { createOAuthStateStore } from './oauth-state-store';
-import type { OAuthStateStore } from './oauth-state.types';
+import { createHash, randomBytes } from 'crypto';
+import { and, eq, gt } from 'drizzle-orm';
+
+import { getDb } from '@/lib/db/runtime';
+import { oauthStateTokens } from '@/lib/db/schema';
+
+const TOKEN_TTL_MS = 10 * 60 * 1000; // 10 minutes
+
+interface OAuthStateStore {
+  issue(params: { authUserId: string; provider?: string }): Promise<string>;
+  consume(params: { stateToken: string }): Promise<string | null>;
+}
+
+function hashToken(token: string): string {
+  return createHash('sha256').update(token).digest('hex');
+}
+
+function generateToken(): string {
+  return randomBytes(32).toString('base64url');
+}
+
+function createOAuthStateStore(): OAuthStateStore {
+  return {
+    async issue({ authUserId, provider }): Promise<string> {
+      const db = getDb();
+      const plainToken = generateToken();
+      const tokenHash = hashToken(plainToken);
+      const expiresAt = new Date(Date.now() + TOKEN_TTL_MS);
+
+      await db.insert(oauthStateTokens).values({
+        stateTokenHash: tokenHash,
+        authUserId,
+        provider: provider ?? null,
+        expiresAt,
+      });
+
+      return plainToken;
+    },
+
+    async consume({ stateToken }): Promise<string | null> {
+      const db = getDb();
+      const tokenHash = hashToken(stateToken);
+      const now = new Date();
+
+      // Atomic delete-and-return: ensures single-use even under concurrent requests
+      const [deleted] = await db
+        .delete(oauthStateTokens)
+        .where(
+          and(
+            eq(oauthStateTokens.stateTokenHash, tokenHash),
+            gt(oauthStateTokens.expiresAt, now)
+          )
+        )
+        .returning({ authUserId: oauthStateTokens.authUserId });
+
+      return deleted?.authUserId ?? null;
+    },
+  };
+}
 
 let defaultStore: OAuthStateStore | null = null;
 
@@ -24,6 +81,3 @@ export async function validateOAuthStateToken(
 ): Promise<string | null> {
   return store.consume({ stateToken });
 }
-
-export { createOAuthStateStore } from './oauth-state-store';
-export type { OAuthStateStore } from './oauth-state.types';
