--- tests/integration/google-calendar-sync.spec.ts
@@ -1,220 +1,0 @@
-import {
-  integrationTokens,
-  learningPlans,
-  modules,
-  tasks,
-  users,
-} from '@/lib/db/schema';
-import { db } from '@/lib/db/service-role';
-import { storeOAuthTokens } from '@/lib/integrations/oauth';
-import { eq } from 'drizzle-orm';
-import { NextRequest } from 'next/server';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { setTestUser } from '../helpers/auth';
-import { ensureUser } from '../helpers/db';
-
-// Mock Auth auth before importing the route
-vi.mock('@/lib/auth/server', () => ({
-  auth: { getSession: vi.fn() },
-}));
-
-// Mock googleapis
-vi.mock('googleapis', () => ({
-  google: {
-    auth: {
-      OAuth2: vi.fn().mockImplementation(() => ({
-        setCredentials: vi.fn(),
-      })),
-    },
-    calendar: vi.fn().mockReturnValue({
-      events: {
-        insert: vi.fn().mockResolvedValue({
-          data: { id: 'event_123', status: 'confirmed' },
-        }),
-      },
-    }),
-  },
-}));
-
-describe.skip('Google Calendar Sync API (temporarily disabled)', () => {
-  let testUserId: string;
-  let testPlanId: string;
-  let testModuleId: string;
-  const authUserId = 'auth_google_sync_test';
-  const createdTaskIds: string[] = [];
-
-  beforeEach(async () => {
-    // Mock Auth auth to return test user
-    const { auth } = await import('@/lib/auth/server');
-    vi.mocked(auth.getSession).mockResolvedValue({
-      data: { user: { id: authUserId } },
-    });
-
-    // Ensure route handlers authenticate as this test user
-    setTestUser(authUserId);
-
-    // Ensure test user
-    testUserId = await ensureUser({
-      authUserId,
-      email: 'google-sync-test@example.com',
-    });
-
-    // Create test plan
-    const [plan] = await db
-      .insert(learningPlans)
-      .values({
-        userId: testUserId,
-        topic: 'Test Google Sync',
-        skillLevel: 'beginner',
-        weeklyHours: 10,
-        learningStyle: 'mixed',
-        generationStatus: 'ready',
-      })
-      .returning();
-    testPlanId = plan.id;
-
-    // Create test module
-    const [module] = await db
-      .insert(modules)
-      .values({
-        planId: testPlanId,
-        order: 1,
-        title: 'Test Module',
-        description: 'Test module description',
-        estimatedMinutes: 120,
-      })
-      .returning();
-    testModuleId = module.id;
-
-    // Create test tasks
-    const createdTasks = await db
-      .insert(tasks)
-      .values([
-        {
-          moduleId: testModuleId,
-          order: 1,
-          title: 'Task 1',
-          description: 'First task',
-          estimatedMinutes: 60,
-        },
-        {
-          moduleId: testModuleId,
-          order: 2,
-          title: 'Task 2',
-          description: 'Second task',
-          estimatedMinutes: 60,
-        },
-      ])
-      .returning({ id: tasks.id });
-
-    createdTaskIds.push(...createdTasks.map((t) => t.id));
-
-    // Store Google Calendar token
-    await storeOAuthTokens({
-      userId: testUserId,
-      provider: 'google_calendar',
-      tokenData: {
-        accessToken: 'test_google_token',
-        refreshToken: 'test_refresh_token',
-        scope: 'https://www.googleapis.com/auth/calendar',
-      },
-    });
-  });
-
-  it('should create calendar events for plan tasks', async () => {
-    const { POST } = await import(
-      '@/app/api/v1/integrations/google-calendar/sync/route'
-    );
-
-    const request = new NextRequest(
-      'http://localhost:3000/api/v1/integrations/google-calendar/sync',
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ planId: testPlanId }),
-      }
-    );
-
-    const response = await POST(request);
-    const data = await response.json();
-
-    expect(response.status).toBe(200);
-    expect(data.success).toBe(true);
-    expect(data.eventsCreated).toBeGreaterThan(0);
-  });
-
-  it('should return 401 if no Google Calendar token found', async () => {
-    const { POST } = await import(
-      '@/app/api/v1/integrations/google-calendar/sync/route'
-    );
-
-    // Delete Google Calendar tokens for this user
-    const { integrationTokens } = await import('@/lib/db/schema');
-    const { and, eq } = await import('drizzle-orm');
-    await db
-      .delete(integrationTokens)
-      .where(
-        and(
-          eq(integrationTokens.userId, testUserId),
-          eq(integrationTokens.provider, 'google_calendar')
-        )
-      );
-
-    const request = new NextRequest(
-      'http://localhost:3000/api/v1/integrations/google-calendar/sync',
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ planId: testPlanId }),
-      }
-    );
-
-    const response = await POST(request);
-
-    expect(response.status).toBe(401);
-  });
-
-  it('should return 400 if planId is missing', async () => {
-    const { POST } = await import(
-      '@/app/api/v1/integrations/google-calendar/sync/route'
-    );
-
-    const request = new NextRequest(
-      'http://localhost:3000/api/v1/integrations/google-calendar/sync',
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({}),
-      }
-    );
-
-    const response = await POST(request);
-
-    expect(response.status).toBe(400);
-  });
-
-  afterEach(async () => {
-    // Clean up test data in reverse order of creation to satisfy foreign key constraints
-    if (createdTaskIds.length > 0) {
-      await db.delete(tasks).where(eq(tasks.moduleId, testModuleId));
-    }
-    if (testModuleId) {
-      await db.delete(modules).where(eq(modules.id, testModuleId));
-    }
-    if (testPlanId) {
-      await db.delete(learningPlans).where(eq(learningPlans.id, testPlanId));
-    }
-    if (testUserId) {
-      await db
-        .delete(integrationTokens)
-        .where(eq(integrationTokens.userId, testUserId));
-      await db.delete(users).where(eq(users.id, testUserId));
-    }
-
-    // Clear mocks
-    vi.clearAllMocks();
-
-    // Reset tracked IDs
-    createdTaskIds.length = 0;
-  });
-});
